assignments/A4/requirements.txt
facenet_pytorch==2.5.3
matplotlib==3.8.4
pandas==2.2.2
Pillow==10.3.0